fix(go/lang): split insecure cookie rule
