+++ rules/go/lang/insecure_cookie.yml
@@ -4,15 +4,10 @@ patterns:
       - variable: COOKIE
         detection: go_lang_insecure_cookie_http_cookie
         scope: cursor_strict
-      - either:
-          - not:
-              variable: COOKIE
-              detection: go_lang_insecure_cookie_secure_true
-              scope: cursor_strict
-          - not:
-              variable: COOKIE
-              detection: go_lang_insecure_cookie_http_only_true
-              scope: cursor_strict
+      - not:
+          variable: COOKIE
+          detection: go_lang_insecure_cookie_true
+          scope: cursor_strict
 auxiliary:
   - id: go_lang_insecure_cookie_cookie_init
     patterns:
@@ -31,16 +26,9 @@ auxiliary:
           - variable: HTTP
             detection: go_lang_insecure_cookie_cookie_init
             scope: cursor
-  - id: go_lang_insecure_cookie_secure_true
+  - id: go_lang_insecure_cookie_true
     patterns:
       - pattern: "$<_>{ Secure: $<TRUE> }"
-        filters:
-          - variable: "TRUE"
-            detection: go_lang_insecure_cookie_true
-            scope: cursor
-  - id: go_lang_insecure_cookie_http_only_true
-    patterns:
-      - pattern: "$<_>{ HttpOnly: $<TRUE> }"
         filters:
           - variable: "TRUE"
             detection: go_lang_insecure_cookie_true
@@ -52,24 +40,25 @@ metadata:
   remediation_message: |
     ## Description
 
-    Cookies without proper security settings can be vulnerable to cross-site scripting (XSS) attacks and potentially provide an avenue for unauthorized access to your application.
+    When set to "true", the "Secure" attribute ensures that a client will only send the cookie to the server when HTTPS is being used.
+    This prevents the cookie from being observed by unauthorized third parties.
 
     ## Remediations
 
     To enhance cookie security and protect against common web exploits:
 
-    ✅ **Use HttpOnly Flag**: Set the `HttpOnly` attribute for cookies to `true`. This prevents client-side scripts from accessing the cookie, reducing the risk of client-side attacks.
+    ✅ Set the `Secure` attribute for cookies to `true`. This prevents the cookie from being observed by unauthorized third parties.
 
     ```go
     http.SetCookie(w, &http.Cookie{
         Name:     "session_token",
         Value:    sessionToken,
-        HttpOnly: true, // Secure the cookie from client-side scripts
-        // Additional flags like Secure, SameSite, etc., should be set as needed.
+        Secure: true,
+        // Additional flags like HttpOnly, SameSite, etc., should be set as needed.
     })
     ```
 
-    ✅ **Additional Cookie Attributes**: Alongside `HttpOnly`, consider setting `Secure`, `SameSite`, and `Domain` attributes to further secure cookies based on your application’s requirements.
+    ✅ Alongside `Secure`, consider setting `HttpOnly`, `SameSite`, and `Domain` attributes to further secure cookies based on your application’s requirements.
 
     ## Resources
 
@@ -78,7 +67,6 @@ metadata:
     - [OWASP Secure Session Management Cheat Sheet](https://cheatsheetseries.owasp.org/cheatsheets/Session_Management_Cheat_Sheet.html)
     - [MDN Web Docs: HttpOnly Cookie Attribute](https://developer.mozilla.org/en-US/docs/Web/HTTP/Cookies#restrict_access_to_cookies)
   cwe_id:
-    - 1004
     - 614
   id: go_lang_insecure_cookie
   documentation_url: https://docs.bearer.com/reference/rules/go_lang_insecure_cookie
